requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
litellm>=1.0.0
nemoguardrails>=0.10.0
